Refactored Update component
