+++ frontend/src/use-cases/update/elements/index.jsx
@@ -1,0 +1,15 @@
+import UserUpdate from "./UserUpdate";
+import UserAddition from "./UserAddition";
+import UserDeletion from "./UserDeletion";
+import GroupUpdate from "./GroupUpdate";
+import GroupAddition from "./GroupAddition";
+import GroupDeletion from "./GroupDeletion";
+
+export {
+  UserUpdate,
+  UserAddition,
+  UserDeletion,
+  GroupUpdate,
+  GroupAddition,
+  GroupDeletion,
+};
