+++ frontend/src/utils/utils.js
@@ -1,0 +1,1 @@
+export const getId = email => email.substr(0, email.search("@"));
